Refactor post move to game component
