+++ src/Game/Board/Board.js
@@ -2,7 +2,6 @@ import React from 'react';
 import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
 import Square from '../Square/Square';
-import { postMove } from '../../client';
 import { getPiece } from '../Piece/Piece';
 import { boardPropType } from '../../logic';
 import { playerPropType } from '../../customPropTypes';
@@ -25,23 +24,11 @@ const Board = ({
   board,
   handleLegalMove,
   handleSelect,
-  fetchGame,
   legalMoves,
   reversed,
   selectedSlot,
-  gameId,
 }) => {
   const getPieceOn = (slot) => getPiece(board.getPiece(slot));
-
-  const getPostMovePayload = (fromSlot, toSlot) => {
-    const { name: player } = activePlayer;
-    const fromPos = board.getRankFile(fromSlot);
-    const toPos = board.getRankFile(toSlot);
-    const piece = board.getPiece(fromSlot);
-    return {
-      player, piece, fromPos, toPos,
-    };
-  };
 
   const selectedCanCapture = (slot) => {
     if (selectedSlot === null) return false;
@@ -57,19 +44,7 @@ const Board = ({
   };
 
   const handleMove = (fromSlot, toSlot) => {
-    if (isLegalMove(fromSlot, toSlot)) {
-      handleLegalMove(fromSlot, toSlot);
-
-      // Post move to server
-      postMove(gameId, getPostMovePayload(fromSlot, toSlot))
-        .then(({ status }) => {
-          if (status !== 201) fetchGame();
-        })
-        .catch(() => {
-          // TODO: display useful error?
-          fetchGame();
-        });
-    }
+    if (isLegalMove(fromSlot, toSlot)) handleLegalMove(board, fromSlot, toSlot);
     handleSelect({ slot: null });
   };
 
@@ -123,11 +98,9 @@ Board.propTypes = {
   board: boardPropType.isRequired,
   handleLegalMove: PropTypes.func.isRequired,
   handleSelect: PropTypes.func.isRequired,
-  fetchGame: PropTypes.func.isRequired,
   legalMoves: PropTypes.arrayOf(PropTypes.arrayOf(PropTypes.number)).isRequired,
   reversed: PropTypes.bool.isRequired,
   selectedSlot: PropTypes.number,
-  gameId: PropTypes.string.isRequired,
 };
 
 Board.defaultProps = {
